Convert tool option size/thickness range to decimal precision

+++ toonz/sources/tnztools/bluredbrush.cpp
@@ -208,7 +208,7 @@ void putOnRaster(const TRaster32P &out, const TRaster32P &in) {
 //
 //=======================================================
 
-BluredBrush::BluredBrush(const TRaster32P &ras, int size,
+BluredBrush::BluredBrush(const TRaster32P &ras, double size,
                          const QRadialGradient &gradient,
                          BrushTipData *brushTip, double spacing,
                          double rotation, bool flipH, bool flipV,
@@ -658,7 +658,7 @@ TRect BluredBrush::getBoundFromPoints(
 //    Blurred Brush implementation
 //*******************************************************************************
 
-RasterBlurredBrush::RasterBlurredBrush(const TRaster32P &ras, int size,
+RasterBlurredBrush::RasterBlurredBrush(const TRaster32P &ras, double size,
                                        const QRadialGradient &gradient,
                                        BrushTipData *brushTip, double spacing,
                                        double rotation, bool flipH, bool flipV,
